[Store][Manticore] Rename `Manticore` to `ManticoreSearch` in namespace and config options
