+++ examples/.env
@@ -180,8 +180,8 @@ POGOCACHE_PASSWORD=symfony
 # Redis (both store and message store)
 REDIS_HOST=localhost
 
-# Manticore (store)
-MANTICORE_HOST=http://127.0.0.1:9308
+# Manticore Search (store)
+MANTICORESEARCH_HOST=http://127.0.0.1:9308
 
 # OpenSearch (store)
 OPENSEARCH_ENDPOINT=http://127.0.0.1:9200
