+++ src/store/src/Bridge/ManticoreSearch/Store.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\AI\Store\Bridge\Manticore;
+namespace Symfony\AI\Store\Bridge\ManticoreSearch;
 
 use Symfony\AI\Platform\Vector\NullVector;
 use Symfony\AI\Platform\Vector\Vector;
